Avoid accessing destroyed contexts in loop
Dispatch globally can cause certain events to close a view with bound context. This causes the destruction of the bound context. If this context has not yet been reached by the _.each loop that triggers the event on all contexts, an attempt will later be made to execute "context.vent.trigger" on a destroyed view and all further processing will abort on a "trying to access vent of undefined" error for that view.